requirements.txt
pandas
openpyxl
streamlit
openai
google-generativeai
matplotlib
seaborn
tabulate
